Three performance fixes: autofocus on settle, dynamic chunk size, mesh cast cleanup
main.rs: autofocus now fires only on the first quiet frame after movement
stops (pending_autofocus flag). During movement the camera rebuilds with
the existing focus_dist, eliminating per-frame refocus jitter.

main.rs: par_chunks_mut chunk size is now computed as
pixels / (num_threads * 4), reducing Rayon task count from ~15k to ~tens
at typical resolutions while preserving 4x oversubscription for load balancing.

mesh.rs: removed no-op `as f32` casts on tobj vertex/normal/texcoord
arrays (tobj already stores them as f32).

+++ src/mesh.rs
@@ -94,17 +94,17 @@ pub fn load_obj(
             let (i0, i1, i2) = (chunk[0] as usize, chunk[1] as usize, chunk[2] as usize);
 
             let v = [
-                Point3::new(pos[3*i0] as f32 * scale, pos[3*i0+1] as f32 * scale, pos[3*i0+2] as f32 * scale),
-                Point3::new(pos[3*i1] as f32 * scale, pos[3*i1+1] as f32 * scale, pos[3*i1+2] as f32 * scale),
-                Point3::new(pos[3*i2] as f32 * scale, pos[3*i2+1] as f32 * scale, pos[3*i2+2] as f32 * scale),
+                Point3::new(pos[3*i0] * scale, pos[3*i0+1] * scale, pos[3*i0+2] * scale),
+                Point3::new(pos[3*i1] * scale, pos[3*i1+1] * scale, pos[3*i1+2] * scale),
+                Point3::new(pos[3*i2] * scale, pos[3*i2+1] * scale, pos[3*i2+2] * scale),
             ];
 
             let max_idx = i0.max(i1).max(i2);
             let n = if nrm.len() >= 3 * (max_idx + 1) {
                 [
-                    Vec3::new(nrm[3*i0] as f32, nrm[3*i0+1] as f32, nrm[3*i0+2] as f32),
-                    Vec3::new(nrm[3*i1] as f32, nrm[3*i1+1] as f32, nrm[3*i1+2] as f32),
-                    Vec3::new(nrm[3*i2] as f32, nrm[3*i2+1] as f32, nrm[3*i2+2] as f32),
+                    Vec3::new(nrm[3*i0], nrm[3*i0+1], nrm[3*i0+2]),
+                    Vec3::new(nrm[3*i1], nrm[3*i1+1], nrm[3*i1+2]),
+                    Vec3::new(nrm[3*i2], nrm[3*i2+1], nrm[3*i2+2]),
                 ]
             } else {
                 let face_n = (v[1] - v[0]).cross(v[2] - v[0]).unit();
@@ -113,9 +113,9 @@ pub fn load_obj(
 
             let uv = if tex.len() >= 2 * (max_idx + 1) {
                 [
-                    (tex[2*i0] as f32, tex[2*i0+1] as f32),
-                    (tex[2*i1] as f32, tex[2*i1+1] as f32),
-                    (tex[2*i2] as f32, tex[2*i2+1] as f32),
+                    (tex[2*i0], tex[2*i0+1]),
+                    (tex[2*i1], tex[2*i1+1]),
+                    (tex[2*i2], tex[2*i2+1]),
                 ]
             } else {
                 [(0.0, 0.0), (1.0, 0.0), (0.0, 1.0)]
